Make the developer wrap the button group items
Instead of creating an observer and moving the button group children into a wrapper, make the developer do it. It will make the DOM structure more clear to the developer, and there won't be any misunderstandings when writing CSS.
Make the `<context-menu>` element extend the `<menu>` element. Use `<menu>` as the button wrapper inside of a submenu